Fix warnings (#200)
* Fix a bunch of warnings

* Fix more warnings. Some warnings aren't obvious how to fix.

* fix a bunch of downcast warnings in DATA

* fix a few more, mostly forward decls

* move forward decls to decomp_/functions.h

* unfinished stuff, doesn't compile but trying to fix warnings. RB_Hazard_HurtDriver's return type is ambiguous atm

* fix signatures and compiles

* renamed RB_Hazard_HurtDriver to DECOMP_

* fixed a bunch more warnings.

* fixed more warnings, re-decomped a couple of functions, begin D231

* fixed more warnings

* fix more warnings

* more warnings

* more warnings

* fix more warnings, changed some signatures, fixed return->return (int) for some LInC functions

* Fixed a bunch more warnings

* fixed a bunch more warnings

* Fixed a bunch more warnings, only have a few tricky ones left.

* Fixed a few more warnings

* fix some pc rebuild warnings

* fixed signature types and some local variable types

* more type fixes

* fixed a bunch more warnings, fixed missing function parameters for some callbacks

* I think that's all?

* fix road hazards and warp orbs

* fix spider web

* Update CrashTeamRacingPC.vcxproj.user

* fix a return type in signature

* fixed a warnings from the merge

* added some comments. I think the credits ghost bug is due to byteBudget/clobbering

* small changes trying to figure out the credits bug

* commit for large test

* fix another warning

* switched file naming convention for BOTS

* Fix file naming convention for more files, rewrite of byteBudget/main.py

* Fixed a bunch more naming conventions

* fix more naming schemes. ModsX.c/ModsMain.c is nonfunctional until further notice

* moving functions to rdata is causing problems, disabled the decomp for MATH_Sin, LibraryOfModels_Store, PROC_SearchForModel

* re-enable 233 cs credits functions

* update bytebudget to display more info

* fix pc port include paths

* fix naming of main functions for bytebudget

* Fix credits bug.

* fix more naming

* fix more naming conventions

* Fixed the rest of the naming in Vehicle. a couple of functions are overbudget.

* fix more naming

* repaired mods

* disable online/defrag

* disable mixrng decomp

* Decomp for MainFrame_HaveAllPads.

* re-enable mixrng getvalue

* fix 230/1 and modify byteBudget to recognize howl naming

* add comments

* add 23x entries for "FULL" bytebudget.

+++ decompile/General/231/231_032_RB_Warpball_SeekDriver.c
@@ -17,7 +17,7 @@ void DECOMP_RB_Warpball_SeekDriver(struct TrackerWeapon* tw,unsigned int param_2
 		&first[param_2];
     
 	while (
-			(d->distanceToFinish_curr <= (cn->distToFinish << 3)) &&
+			(d->distanceToFinish_curr <= (unsigned int)(cn->distToFinish << 3)) &&
            
 			// node is not first node
 			(cn != first)
